% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\docType{data}
\name{allRegions}
\alias{allRegions}
\title{Polygon of all regions}
\format{\code{SpatialPolygons} object}
\usage{
allRegions
}
\description{
All regions except the non-regional ocean converted into a single \code{SpatialPolygons}
object. Regions are modified from the region masks provided by the National Snow and Ice Data
Center (NSIDC)
}
\examples{
data(allRegions)
plot(allRegions)
}
\references{
Region Mask:
National Snow and Ice Data Center, 2017: Region mask for the northern hemisphere.
\url{http://nsidc.org/data/polar-stereo/tools_masks.html}.
}
\keyword{datasets}
