% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfIntersections.R
\name{anyIntersect}
\alias{anyIntersect}
\title{Check if a line has intersecting segments}
\usage{
anyIntersect(line)
}
\arguments{
\item{line}{matrix of coordinates corresponding to the line of interest}
}
\value{
list where \code{list$any} is a boolean indicating if there are any intersections
and \code{list$val} is an index corresponding to the first intersection found
}
\description{
Determine if there are any intersecting line segments in a matrix of coordinates representing a line
}
\examples{
checkResults <- anyIntersect(currSecEx)
checkResults$any #true/false
checkResults$val #indices of first intersection found
}
