% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\docType{data}
\name{bgWater}
\alias{bgWater}
\title{Polygon of the non-regional ocean}
\format{\code{SpatialPolygons} object}
\usage{
bgWater
}
\description{
The non-regional ocean converted into a single \code{SpatialPolygons}
object. Thee boundaries of the non-regional ocean were defined by modifying the region masks provided
by the National Snow and Ice Data  Center (NSIDC).
}
\examples{
data(bgWater)
plot(bgWater)
}
\references{
Region Mask:
National Snow and Ice Data Center, 2017: Region mask for the northern hemisphere.
\url{http://nsidc.org/data/polar-stereo/tools_masks.html}.
}
\keyword{datasets}
