% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{fixedMapping}
\alias{fixedMapping}
\title{Map any region with a fixed set of lines}
\usage{
fixedMapping(ice, region, fixLines, startCoords, plotting = FALSE)
}
\arguments{
\item{ice}{SpatialPolygon represetnting the area covered in ice}

\item{region}{SpatialPolygon giving the region of interest}

\item{fixLines}{Set of SpatialLines on which the values will be mapped}

\item{startCoords}{matrix of coordinates from which the lengths will be mapped}

\item{plotting}{boolean indicating if the vectors should be mapped}
}
\description{
Find the mapping for any region
}
\details{
Note that this function differs from other mapping functions
in that it works only with a fixed set of lines
}
