% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getCoords}
\alias{getCoords}
\title{Extract coordinates from a spatial object of lines and points}
\usage{
getCoords(myPoints)
}
\arguments{
\item{myPoints}{spatial object of type \code{SpatialCollections}, \code{SpatialPoints}, or \code{SpatialLines}}
}
\value{
n x 2 matrix of coordinates
}
\description{
Get coordinates from a spatial object of lines and points. There is no ordering of points returned.
 Note: This differs from \code{extractCoords} in that the ordering of the points is NOT considered.
}
\examples{
#Load sample line
exampleLine <- as(rmHoles(bgWater[2]), "SpatialLines")
getCoords(exampleLine)
}
