% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getDist}
\alias{getDist}
\title{Find euclidean distance}
\usage{
getDist(p1, p2)
}
\arguments{
\item{p1}{first point,  x and y coordinate pair}

\item{p2}{second point,  x and y coordinate pair}
}
\value{
distance value
}
\description{
Finds the euclidean distance between two points (ignoring projection)
}
\examples{
getDist(c(1, 2), c(3, 4))
}
