% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probContourInfo.R
\name{getXInfo}
\alias{getXInfo}
\title{Get info on hidden values}
\usage{
getXInfo(yMat, distList)
}
\arguments{
\item{yMat}{matrix of observed distances (dimension number of vectors x number of years)}

\item{distList}{list of boundary points for each of the vectors (length: number of vectors)}
}
