% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings.R
\name{intLine}
\alias{intLine}
\title{Space points along a line}
\usage{
intLine(predL, obsL, plotting = FALSE)
}
\arguments{
\item{predL}{predicted line (n1 x 2 matrix of coordinates)}

\item{obsL}{predicted line (n2 x 2 matrix of coordinates)}

\item{plotting}{boolean indicating whether maps should be plotted}
}
\value{
n x 2 matrix of evenly-spaced coordinates
}
\description{
The function evenly spaces the number of points that are on one line, \code{predL}, on a
different line, \code{obsL}
}
\examples{
lineSpace <- intLine(predLEx, obsLEx, plotting = TRUE)
}
