% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{keepLine}
\alias{keepLine}
\title{Keep only spatial lines}
\usage{
keepLine(myPoly)
}
\arguments{
\item{myPoly}{\code{SpatialCollections}, \code{SpatialPolygons}, \code{SpatialPoints}, or \code{SpatialLines} object}
}
\value{
\code{SpatialPolygons} object
}
\description{
Keep only \code{SpatialLines} from a spatial object.
}
\examples{
par(mfrow = c(1, 2))
plot(spatialCollEx, col = "blue", main = "Spatial Collections Object")
lineOnly <- keepLine(spatialCollEx)
plot(lineOnly, col = "blue", main = "Spatial Line Only")
}
