% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{keepPoly}
\alias{keepPoly}
\title{Keep only spatial polygons}
\usage{
keepPoly(myPoly)
}
\arguments{
\item{myPoly}{\code{SpatialCollections}, \code{SpatialPolygons}, \code{SpatialPoints}, or \code{SpatialLines} object}
}
\value{
\code{SpatialPolygons} object
}
\description{
Keep only \code{SpatialPolygons} from a spatial object.
}
\examples{
par(mfrow = c(1, 2))
plot(spatialCollEx, col = "blue", main = "Spatial Collections Object")
polyOnly <- keepPoly(spatialCollEx)
plot(polyOnly, col = "blue", main = "Spatial Polygon Only")
}
