% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{obsFeb19811982}
\alias{obsFeb19811982}
\title{Observed sea ice February 1981-1982}
\format{array of dimension of 2 x 12 x 304 x 448 (year x month x longitude x latitude)y}
\usage{
obsFeb19811982
}
\description{
The object \code{observed} is an array obtained from the function \code{readMonthlyBS}
for startYear = 1981 and endYear = 1982. It gives the observed sea ice
concentrations arranged in an array of dimension of year x month x lon x lat. The observations are from the monthly sea ice concentration
obtained from the National Aeronautics and Space Administration (NASA) satellites Nimbus-7
SMMR and DMSP SSM/I-SSMIS and processed by the bootstrap algorithm. The results
are distributed by the National Snow and Ice Data Center (Comiso 2000, updated 2015).
}
\examples{
data(obsFeb19811982)
dim(obsFeb19811982)
}
\references{
Bootstrap sea ice concentration:

Comiso, J., 2000, updated 2015: Bootstrap sea ice concentrations from Nimbus-7 SMMR and
DMSP SSM/I-SSMIS. version 2. \url{http://nsidc.org/data/nsidc-0079}
}
\keyword{datasets}
