% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storedResults.R
\docType{data}
\name{obsFeb2012}
\alias{obsFeb2012}
\title{Observed sea ice February 2012}
\format{array of dimension of 2 years x 12 months  x 304 longitudes x 448 latitudes}
\usage{
obsFeb2012
}
\description{
The object \code{observed} is an array obtained from the \code{readMonthlyBS} function
for startYear = 2012 and endYear = 2012. It gives the observed sea ice
concentrations arranged in an array of dimension of year x month x lon x lat
}
\examples{
data(obsFeb2012)
dim(obsFeb2012)
}
\references{
Bootstrap sea ice concentration:
Comiso, J., 2000, updated 2015: Bootstrap sea ice concentrations from Nimbus-7 SMMR and
DMSP SSM/I-SSMIS. version 2. \url{http://nsidc.org/data/nsidc-0079}
}
\keyword{datasets}
