% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionInformation.R
\docType{data}
\name{regionInfo}
\alias{regionInfo}
\title{Information on the regions}
\format{This region information list is a list six objects: \code{regions}, \code{lines}, \code{out}, \code{centRegion}, \code{centLines}, and
\code{centFrom}. The first three objects are ordered lists giving information about each of the regions that will be mapped outside the
Central Arctic region. The variable \code{regions} gives \code{SpatialPolygons} objects for the corresponding regions and the variable \code{lines}
 gives the \code{SpatialLines} objects for the corresponding fixed lines. The variable \code{out} gives \code{SpatialPolygons} objects that are outside the corresponding regions,
 but that border the fixed lines. These are used when building new polygons to determine if points are being mapped outside the region of interest.
 The variable \code{centRegions} is the \code{SpatialPolygons} object corresponding to the central Arctic region, \code{centLines} is the \code{SpatialLines} object for the fixed line,}
\usage{
regionInfo
}
\description{
Information on the regions
}
\examples{
data(regionInfo)
names(regionInfo)
}
\references{
The regions in this object have been modified from the following region mask:

National Snow and Ice Data Center, 2017: Region mask for the northern hemisphere
\url{http://nsidc.org/data/polar-stereo/tools_masks.html}.
}
\keyword{datasets}
